src/requirements.txt
openai
PyPDF2
langchain
langchain_openai
python-decouple
gradio
bs4